feat(filter): made filter as component and put on franchises
