+++ src/lib/i18n/locales/en-US/common.json
@@ -88,6 +88,7 @@
   },
   "minEpisodes": "Min Episodes",
   "minChapters": "Min Chapters",
+  "minPages": "Min Pages",
   "mostPopular": "Most Popular",
   "recentlyReleased": "Recently Released",
   "comingSoon": "Coming Soon",
